+++ server/proxy/transport_windows.go
@@ -1,0 +1,11 @@
+// +build windows
+
+package proxy
+
+import (
+	"github.com/cnlh/nps/lib/conn"
+)
+
+func HandleTrans(c *conn.Conn, s *TunnelModeServer) error {
+	return nil
+}
